+++ sdk/communication/Azure.Communication.Rooms/CHANGELOG.md
@@ -9,6 +9,13 @@
 ### Bugs Fixed
 
 ### Other Changes
+## 1.1.0-beta.1 (2023-09-19)
+
+### Features Added
+- Added Support for PSTN DialOut Capability in Azure Communication Services Rooms
+
+### Added
+- Added new overloads for CreateRoom (CreateRoomAsync) and UpdateRooms (UpdateRoomsAsync) with options : RoomCreateOptions and RoomUpdateOptions. 
 
 ## 1.0.0 (2023-06-12)
 General Availability version of the Azure Communication Services Rooms .NET SDK.